feat(onboarding): celebrate warehouse connect, drop the project-ready screen
Replace the interstitial "We're getting your project ready" screen with a
confetti burst at the moment of success. On a successful Test & save in the
warehouse connection form, fire confetti from the button, then redirect
straight to the project home.

The readiness screen was pure theatre — a fixed 3-step animation over a
single schema/catalog fetch with no real backend work — so the flourish at
the point of success conveys the same delight without the fake progress page.

- CreateProjectConnection: fire-once confetti (guarded against the success
  effect re-running on every job-status poll), originating from the submit
  button, behind a celebrateOnSuccess prop
- OnboardingDataSource: redirect to /projects/:uuid/home + celebrateOnSuccess
- remove the /onboarding/project-ready route, page, styles, and the now-unused
  ONBOARDING_PROJECT_READY page-view event

Refs: PROD-9284
