+++ packages/frontend/src/Routes.tsx
@@ -887,28 +887,6 @@ const PRIVATE_ROUTES: RouteObject[] = [
                 },
             },
             {
-                path: '/onboarding/project-ready/:projectUuid',
-                handle: { hideAILauncher: true },
-                lazy: async () => {
-                    const OnboardingProjectReady = await loadLazyRouteDefault(
-                        './pages/OnboardingProjectReady',
-                        () => import('./pages/OnboardingProjectReady'),
-                    );
-                    return {
-                        Component: () => (
-                            <>
-                                <NavBar />
-                                <TrackPage
-                                    name={PageName.ONBOARDING_PROJECT_READY}
-                                >
-                                    <OnboardingProjectReady />
-                                </TrackPage>
-                            </>
-                        ),
-                    };
-                },
-            },
-            {
                 path: '/createProject/:method?',
                 lazy: async () => {
                     const CreateProject = await loadLazyRouteDefault(
