+++ packages/frontend/src/components/ProjectConnection/ProjectConnectFlow/ConnectManually/ConnectManuallyStep2.tsx
@@ -13,6 +13,7 @@ interface ConnectManuallyStep2Props {
     warehouseOnly?: boolean;
     onBack: () => void;
     successRedirect?: (projectUuid: string) => string;
+    celebrateOnSuccess?: boolean;
 }
 
 const ConnectManuallyStep2: FC<ConnectManuallyStep2Props> = ({
@@ -21,6 +22,7 @@ const ConnectManuallyStep2: FC<ConnectManuallyStep2Props> = ({
     warehouseOnly = false,
     onBack,
     successRedirect,
+    celebrateOnSuccess = false,
 }) => {
     return (
         <>
@@ -43,6 +45,7 @@ const ConnectManuallyStep2: FC<ConnectManuallyStep2Props> = ({
                     selectedWarehouse={selectedWarehouse}
                     warehouseOnly={warehouseOnly}
                     successRedirect={successRedirect}
+                    celebrateOnSuccess={celebrateOnSuccess}
                 />
             </Stack>
         </>
